requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.11.0
idna==3.7
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
multidict==6.0.4
nest_asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
pyzmq==25.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
scipy==1.13.1
setuptools==72.1.0
six==1.16.0
stack-data==0.6.2
tensorboard==2.12.1
tensorboard_data_server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.1.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.35.1
wrapt==1.14.1
yarl==1.9.3
zipp==3.20.1
